pyproject.toml
[project]
name = "bluesky-sample"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "okmtdev", email = "okmtdev@gmail.com" }
]
dependencies = [
    "atproto>=0.0.45",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.rye.scripts]
main = { cmd = "python3 src/bluesky_sample/main.py" }
